Add new test-sessions module to do testing of various session implementations. The test-session-common module was created based on the terracotta session test suite checked into jetty-codehaus. I've genericized these tests so that they may be applied to the various session clustering solutions.
Changes to JDBCSessionIdManager with this checkin were to increase the length of the varchar storing the session id from 60 to 120 char, and also to reduce the amount of time we held a lock over the session id map.

Changes to the JDBCSessionManager with this checkin were to decrease the amount of time we held locks, and when retrieving a session from the database, checking if the session had expired before loading it.

Changes to the AbstractSessionManager with this checkin were to ensure that the SessionListeners were only called if the session was removed (we were calling them anyway).


git-svn-id: http://dev.eclipse.org/svnroot/rt/org.eclipse.jetty/jetty/trunk@1213 7e9141cc-0065-0410-87d8-b60c137991c4