JETTY-776
Remove LightLoad test as this is not applicable for the JDBC solution. It only works with terracotta because terracotta takes out a cluster-wide lock when the session is entered on any node. This means you can bounce the session around nodes. This does not work with the JDBC session implementation as it is not feasible to do a cluster-wide lock on particular row on the sessions table.

Add a method on JDBCSessionManager that would allow a subclass to send a message to all other nodes whenever the session is being written to the database which would cause the other nodes to invalidate that session in their local caches and re-read from the database.


git-svn-id: http://dev.eclipse.org/svnroot/rt/org.eclipse.jetty/jetty/trunk@1218 7e9141cc-0065-0410-87d8-b60c137991c4